move authtoken package into auth package
