--- pkg/services/auth/auth.go
@@ -1,1 +1,0 @@
-package auth
