+++ pkg/services/auth/token_cleanup_test.go
@@ -1,4 +1,4 @@
-package authtoken
+package auth
 
 import (
 	"fmt"
